#3 - migrate database
