+++ .env
@@ -3,4 +3,4 @@ SECRET_KEY=vananhtran0421
 FLASK_ENV=development
 FLASK_APP=application.py
 # Database configuration
-DATABASE_URL=mysql+pymysql://root:12345678@localhost/flask-new?charset=utf8mb4
+DATABASE_URL=mysql+pymysql://root:12345678@localhost/python-flask-news?charset=utf8mb4
