+++ app/__init__.py
@@ -2,9 +2,16 @@ import logging
 from flask import Flask
 from config import Config
 from logging.handlers import TimedRotatingFileHandler
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
 
+# Khởi tạo ứng dụng Flask
 app = Flask(__name__, static_folder='static', static_url_path='/static')
 app.config.from_object(Config)
+
+# Khởi tạo SQLAlchemy và Migrate
+db = SQLAlchemy(app)
+migrate = Migrate(app, db)
 
 # Configuration for media files
 app.config['MEDIA_FOLDER'] = 'media'
@@ -24,4 +31,7 @@ from app.main import main as main_blueprint
 app.register_blueprint(main_blueprint)
 
 from app.admin import admin as admin_blueprint
-app.register_blueprint(admin_blueprint)
+app.register_blueprint(admin_blueprint, url_prefix='/admin')
+
+from app.models import base, categories, posts  # Import models để chúng có thể được phát hiện bởi Flask-Migrate
+
