+++ app/admin/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 admin = Blueprint('admin', __name__)
 
-from . import routes
+from . import routes, models
