--- app/app.py
@@ -1,17 +1,0 @@
-# from flask import render_template, app
-# from app import create_app
-#
-# @app.route('/')
-# def home():
-#     return render_template('index.html')
-#
-# @app.route('/contact')
-# def contact():
-#     return render_template('contact.html')
-#
-#
-# app = create_app()
-#
-# if __name__ == '__main__':
-#     from app.admin import *
-#     app.run(debug=True)
