+++ app/models/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseModel
+from .categories import Category
+from .posts  import Post
