+++ app/models/common_fields.py
@@ -1,0 +1,3 @@
+from sqlalchemy import Column, String, Text, Boolean, DateTime, ForeignKey
+from sqlalchemy.orm import relationship
+from datetime import datetime
